+++ src/app/app.component.html
@@ -1,110 +1,8 @@
-<div style="height: 100%; width: 100%;padding: 10px;">
-  <div>
-      <div class="row action-panel filter-panel-padding">
-          <div>
-              <mat-card class="form-inline" style="padding-bottom: 10px;padding-top: 10px;">
-                  <div class="row" style="margin-bottom: 10px;padding-left: 15px;">
-                      <div class="filter-panel-alignment">
-                          <div>
-                              <div class="sub-textarea-div" style="margin-right:10px">
-                                  <textarea class="form-control" (input)="onValueChange($event)" name="sql_query" id="sql_query_id" placeholder="Enter SQL Query here" rows="5" cols="120"></textarea>
-                              </div>
-                          </div>
-                          <div>
-                              <span class="alertvmsg" *ngIf="isqueryempty">
-                                  Please {{defaultQuery}} sql query
-                              </span>
-                          </div>
-                      </div>
-
-                      <div class="filter-panel-alignment" style="margin-top:40px; margin-left:10px;">
-                          <div></div>
-                          <p-button label="Execute" id='execute-btn' type='button' (click)="executeQueryButton()"></p-button>
-                          <p-button label="Close Snapshots" [loading]="loading" (click)="getCloseTreeGridData()"></p-button>
-                      </div>
-                  </div>
-              </mat-card>
-          </div>
-      </div>
-      <div>
-        <span class="notemsg">
-            {{noteMessage}}
-        </span>
-      </div>
-
-
-      
-  <!-- <p-treeTable [value]="file" [columns]="cols" selectionMode="single" [(selection)]="selectedNode" dataKey="name" (onNodeSelect)="nodeSelect($event)" (onNodeUnselect)="nodeUnselect($event)" [scrollable]="true" [tableStyle]="{'min-width':'50rem'}">
-    <ng-template pTemplate="header" let-columns>
-        <tr>
-            <th *ngFor="let col of columns">
-                {{ col.header }}
-            </th>
-        </tr>
-    </ng-template>
-    <ng-template pTemplate="body" let-rowNode let-rowData="rowData" let-columns="columns">
-        <tr [ttRow]="rowNode" [ttRow]="rowNode" [ttSelectableRow]="rowNode">
-            <td *ngFor="let col of columns; let i = index">
-                <p-treeTableToggler [rowNode]="rowNode" *ngIf="i === 0"></p-treeTableToggler>
-                {{ rowData[col.field] }}
-            </td>
-        </tr>
-    </ng-template>
-  </p-treeTable> -->
-  <div *ngIf="isUserQueryInterface">
-    <p-toast></p-toast>
-    <p-treeTable [value]="data" [columns]="columns" [scrollable]="true" [paginator]="true" [rows]="10" selectionMode="single" [(selection)]="selectedNode" dataKey="name" (onNodeSelect)="nodeSelect($event)" (onNodeUnselect)="nodeUnselect($event)" [tableStyle]="{'min-width':'50rem'}">
-      <ng-template pTemplate="caption"> Query Response </ng-template>
-      <ng-template pTemplate="header" let-columns>
-        <!-- Define the table column headers -->
-        <tr>
-          <th *ngFor="let column of columns" [ttSortableColumn]="column.field">{{ column.field }}
-          <!-- Sort table -->
-            <p-treeTableSortIcon [field]="column.field"></p-treeTableSortIcon>
-          </th>
-        </tr>
-      </ng-template>
-      <ng-template pTemplate="body" let-rowNode let-rowData="rowData" let-columns="columns">
-        <!-- Define the table rows and cells -->
-        <tr [ttRow]="rowNode" [ttSelectableRow]="rowNode">
-          <td *ngFor="let column of columns; let i = index">
-            <p-treeTableToggler [rowNode]="rowNode" *ngIf="i === 0"></p-treeTableToggler>
-            {{ rowData[column.field] }}
-          </td>
-        </tr>
-      </ng-template>
-      <!-- <ng-template pTemplate="summary">
-        <div style="text-align:left">
-            <p-button icon="pi pi-refresh" label="Reload"></p-button>
-        </div>
-      </ng-template> -->
-      <ng-template pTemplate="paginatorleft">
-        <p-button icon="pi pi-refresh" styleClass="p-button-text"></p-button>
-      </ng-template>
-      <ng-template pTemplate="paginatorright">
-          <p-button icon="pi pi-download" styleClass="p-button-text"></p-button>
-      </ng-template>
-    </p-treeTable>
-  </div>
-  
-  <div *ngIf="isCloseSnapshots">
-    <p-treeTable [value]="data">
-      <ng-template pTemplate="header" let-rowNode>
-        <!-- Define the table column headers -->
-        <tr>
-          <th *ngFor="let column of closeGridColumns">{{ column.header }}</th>
-        </tr>
-      </ng-template>
-      <ng-template pTemplate="body" let-rowNode let-rowData="rowData">
-        <!-- Define the table rows and cells -->
-        <tr>
-          <td *ngFor="let column of closeGridColumns">{{rowData[column.field]}}</td>
-        </tr>
-      </ng-template>
-    </p-treeTable>
-  </div>
-      
+<!-- Header Toolbar Start -->
+<app-home></app-home>
+<div style="padding-left: 10px;padding-right: 10px;padding-top: 15px;">
+  <div class="text-font-format">
+    <router-outlet></router-outlet>
   </div>
 </div>
-
