+++ src/app/components/close-list/close-list.component.html
@@ -1,1 +1,4 @@
-<p>close-list works!</p>
+
+<div class="card flex justify-content-center">
+    <p-button label="Close List Grid" [loading]="loading" (click)="getCloseListGrid()"></p-button>
+</div>
