+++ src/app/components/home/home.component.html
@@ -1,0 +1,5 @@
+
+<!-- <router-outlet></router-outlet> -->
+<div class="card">
+    <p-tabMenu [model]="items" [activeItem]="activeItem" (activeItemChange)="onActiveItemChange($event)"></p-tabMenu>
+</div>
