+++ src/app/components/tree-table/tree-table.component.scss
@@ -1,0 +1,3 @@
+#execute-btn{
+    margin-right: 10px;
+}
